requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.35.0
datasets==2.14.0
tokenizers==0.14.0
accelerate==0.24.0
wandb==0.16.0
numpy==1.24.0
pyyaml==6.0
pytest==7.4.0
black==23.11.0
flake8==6.1.0
